Tedious as it was, the language feature has now been reimplemented. I did it a slightly different way than before to try out different things

+++ src/main/java/com/example/sakilaAPI/dto/input/FilmInput.java
@@ -30,6 +30,6 @@ public class FilmInput {
     private String description;
 
     @NotNull(groups = {ValidationGroup.Create.class}, message = "Language ID cannot be null")
-    private Long languageId;
+    private Short languageId;
 
 }
